fix: 🐛 cleanups before prod
