+++ server/src/internal/analytics/actions/eventActions.ts
@@ -1,9 +1,9 @@
-import { aggregate } from "./aggregate.js";
+import { aggregate } from "./eventValidationUtils.js";
 import { getCountAndSum } from "./getCountAndSum.js";
 import { getEventById } from "./getEventById.js";
 import { getTopEventNames } from "./getTopEventNames.js";
 import { listEventNames } from "./listEventNames.js";
-import { listEventsForApi } from "./listEventsForApi.js";
+import { listEvents } from "./listEvents.js";
 import { listRawEvents } from "./listRawEvents.js";
 
 export const eventActions = {
@@ -12,6 +12,6 @@ export const eventActions = {
 	getEventById,
 	getTopEventNames,
 	listEventNames,
-	listEventsForApi,
+	listEvents,
 	listRawEvents,
 } as const;
