+++ server/src/internal/analytics/actions/listEvents.ts
@@ -1,21 +1,10 @@
 import type { ApiEventsListItem } from "@autumn/shared";
+import { epochToDateTime } from "@autumn/shared/api/common/epochUtils";
 import { getTinybirdPipes } from "@/external/tinybird/initTinybird.js";
 import type { AutumnContext } from "@/honoUtils/HonoEnv.js";
 
-/** Converts epoch ms to ClickHouse DateTime string format */
-const epochToDateTime = (epochMs: number): string => {
-	const date = new Date(epochMs);
-	const year = date.getUTCFullYear();
-	const month = String(date.getUTCMonth() + 1).padStart(2, "0");
-	const day = String(date.getUTCDate()).padStart(2, "0");
-	const hours = String(date.getUTCHours()).padStart(2, "0");
-	const minutes = String(date.getUTCMinutes()).padStart(2, "0");
-	const seconds = String(date.getUTCSeconds()).padStart(2, "0");
-	return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-};
-
 /** Lists events for the external API with offset-based pagination */
-export const listEventsForApi = async ({
+export const listEvents = async ({
 	ctx,
 	params,
 }: {
